# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sugar Production Planning API
  version: "1.0.0"
  description: |
    Multi-company production planning versus actual.

    **One login, many companies.** The access token carries no company. Every
    company-dependent call names the company explicitly — as a path parameter,
    a `companyId` query parameter or the `X-Company-Id` header — and the server
    re-validates that the caller is entitled to it. The client-supplied company
    is a request parameter, never an assertion of rights.

    Because there is no global company context, two browser tabs can operate on
    two companies at the same time without interfering.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Sign in
      description: |
        Returns the token pair together with every company the user is
        authorised for and the permissions that apply in each. This response
        is what fills the company drop-down of each screen.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Signed in
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { description: Too many attempts }
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Exchange a refresh token
      description: |
        Rotates the refresh token: the presented one is revoked and linked to
        its successor, so replaying a stolen token after a legitimate refresh
        fails and is visible in the chain.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change the caller's password
      description: Existing sessions are revoked, so a compromised session cannot survive the change.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 12 }
      responses:
        '204': { description: Password changed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/BusinessRule' }

  # --- master data --------------------------------------------------------
  /materials:
    get:
      tags: [Master data]
      summary: List materials (cross-company master data)
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Search'
      responses:
        '200': { $ref: '#/components/responses/MaterialList' }
    post:
      tags: [Master data]
      summary: Create a material
      parameters: [{ $ref: '#/components/parameters/CompanyIdQuery' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaterialRequest' }
      responses:
        '201': { $ref: '#/components/responses/Material' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /materials/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
      - $ref: '#/components/parameters/CompanyIdQuery'
    get:
      tags: [Master data]
      summary: Read one material
      responses:
        '200': { $ref: '#/components/responses/Material' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Master data]
      summary: Change a material
      description: The request must carry the version it read; a stale version is a 409.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MaterialRequest' }
      responses:
        '200': { $ref: '#/components/responses/Material' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Master data]
      summary: Deactivate a material
      description: Master data is deactivated, never physically deleted.
      parameters:
        - name: version
          in: query
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deactivated }
        '409': { $ref: '#/components/responses/Conflict' }
  /companies/{companyId}/materials:
    parameters: [{ $ref: '#/components/parameters/CompanyIdPath' }]
    get:
      tags: [Master data]
      summary: Materials with their relevance flags for this company
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
    put:
      tags: [Master data]
      summary: Set the relevance flags of one material in this company
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompanyMaterialRequest' }
      responses:
        '200': { $ref: '#/components/responses/GenericObject' }
  /companies/{companyId}/warehouses:
    parameters: [{ $ref: '#/components/parameters/CompanyIdPath' }]
    get:
      tags: [Master data]
      summary: List warehouses, tanks and silos
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
    post:
      tags: [Master data]
      summary: Create a storage location
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseRequest' }
      responses:
        '201': { $ref: '#/components/responses/GenericObject' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /companies/{companyId}/warehouses/{id}:
    parameters:
      - $ref: '#/components/parameters/CompanyIdPath'
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Master data]
      summary: Read one storage location with its allowed materials
      responses:
        '200': { $ref: '#/components/responses/GenericObject' }
    put:
      tags: [Master data]
      summary: Change a storage location
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseRequest' }
      responses:
        '200': { $ref: '#/components/responses/GenericObject' }
    delete:
      tags: [Master data]
      summary: Deactivate a storage location
      parameters:
        - name: version
          in: query
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deactivated }
  /companies/{companyId}/production-lines:
    parameters: [{ $ref: '#/components/parameters/CompanyIdPath' }]
    get:
      tags: [Master data]
      summary: List production lines
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
    post:
      tags: [Master data]
      summary: Create a production line
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        '201': { $ref: '#/components/responses/GenericObject' }
  /companies/{companyId}/seasons:
    parameters: [{ $ref: '#/components/parameters/CompanyIdPath' }]
    get:
      tags: [Master data]
      summary: List seasons
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
    post:
      tags: [Master data]
      summary: Create a season
      description: Seasons of one company must not overlap; an overlap is `E-VAL-013`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SeasonRequest' }
      responses:
        '201': { $ref: '#/components/responses/GenericObject' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /processes:
    get:
      tags: [Master data]
      summary: List production processes
      parameters: [{ $ref: '#/components/parameters/CompanyIdQuery' }]
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
  /processes/{id}/materials:
    put:
      tags: [Master data]
      summary: Declare the input and output materials of a process
      description: |
        This is what drives the production routing rules. A material flagged
        `requiresConditioning` must reach its finished-goods warehouse through
        the condition silo; adding a new route is a master-data change, not a
        code change.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/CompanyIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/ProcessMaterialRequest' }
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
  /movement-types:
    get:
      tags: [Master data]
      summary: List movement types
      description: |
        `direction` is the contract: an inventory quantity is always positive
        and the movement type decides whether it adds to or subtracts from
        stock.
      parameters: [{ $ref: '#/components/parameters/CompanyIdQuery' }]
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
  /uoms:
    get:
      tags: [Master data]
      summary: List units of measure
      parameters: [{ $ref: '#/components/parameters/CompanyIdQuery' }]
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
  /packaging-types:
    get:
      tags: [Master data]
      summary: List packaging types
      parameters: [{ $ref: '#/components/parameters/CompanyIdQuery' }]
      responses:
        '200': { $ref: '#/components/responses/GenericList' }

  # --- planning -----------------------------------------------------------
  /companies/{companyId}/seasons/{seasonId}/planning-versions:
    parameters:
      - $ref: '#/components/parameters/CompanyIdPath'
      - name: seasonId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Planning]
      summary: List the planning versions of a season
      responses:
        '200': { $ref: '#/components/responses/GenericList' }
    post:
      tags: [Planning]
      summary: Create a planning version
      description: |
        Version numbering is open-ended. Omit `versionNo` to let the server
        allocate the next free number.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlanningVersionRequest' }
      responses:
        '201': { $ref: '#/components/responses/PlanningVersion' }
        '409': { $ref: '#/components/responses/Conflict' }
  /planning-versions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
      - $ref: '#/components/parameters/CompanyIdQuery'
    get:
      tags: [Planning]
      summary: Read a planning version
      responses:
        '200': { $ref: '#/components/responses/PlanningVersion' }
    put:
      tags: [Planning]
      summary: Change the descriptive attributes of a version
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlanningVersionRequest' }
      responses:
        '200': { $ref: '#/components/responses/PlanningVersion' }
  /planning-versions/{id}/copy:
    post:
      tags: [Planning]
      summary: Copy a version into a new, independent one
      description: |
        The copy is a deep snapshot: every header and item is duplicated and
        nothing is shared with the source. Editing the copy can never change
        the original — the only link is the informational
        `copiedFromVersionId`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/CompanyIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [versionName]
              properties:
                newVersionNo: { type: integer, minimum: 1 }
                versionName: { type: string }
                description: { type: string, nullable: true }
      responses:
        '201': { $ref: '#/components/responses/PlanningVersion' }
        '409': { $ref: '#/components/responses/Conflict' }
  /planning-versions/{id}/submit:
    post:
      tags: [Planning]
      summary: Submit a draft for approval
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/CompanyIdQuery'
      responses:
        '200': { $ref: '#/components/responses/PlanningVersion' }
        '409': { $ref: '#/components/responses/Conflict' }
  /planning-versions/{id}/approve:
    post:
      tags: [Planning]
      summary: Approve a submitted version
      description: |
        With the four-eyes rule enabled the approver must be somebody other
        than the creator, which is `E-PLAN-010`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/CompanyIdQuery'
      responses:
        '200': { $ref: '#/components/responses/PlanningVersion' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /planning-versions/{id}/lock:
    post:
      tags: [Planning]
      summary: Lock an approved version
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/CompanyIdQuery'
      responses:
        '200': { $ref: '#/components/responses/PlanningVersion' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The access token identifies the user and nothing else. It carries no
        company, which is why every company-dependent call must name one.
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...
  schemas:
    Envelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: {}
        meta: { $ref: '#/components/schemas/Meta' }
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
...
    LoginResponse:
      type: object
      properties:
        tokens: { $ref: '#/components/schemas/TokenPair' }
        user: { $ref: '#/components/schemas/User' }
        companies:
          type: array
          items: { $ref: '#/components/schemas/CompanyAccess' }
        mustChangePassword: { type: boolean }
...